interceptor

+++ Core/NlayeredApp.Domain/Entities/Product.cs
@@ -14,7 +14,7 @@ namespace NlayeredApp.Domain.Entities
         {
         }
 
-        public Product(string name, int stock, long price)
+        public Product(string name, int stock, float price)
         {
             Name = name;
             Stock = stock;
@@ -23,7 +23,7 @@ namespace NlayeredApp.Domain.Entities
 
         public string Name { get; set; }
         public int Stock { get; set; }
-        public long Price { get; set; }
+        public float Price { get; set; }
 
         public ICollection<Order> Orders { get; set; }
